Rule collision logic
Keep in mind that the non-explicit rules are allowed to collide with the explicit ones because the explicit ones always have priority. So the collision is most interesting for the explicit rules (there it's much simpler as well). Within the non-explicit rules it's also interesting because we need to avoid running into a situation where two rules of the same priority (first explicitness, then size of from) match a tag set. Unfortunately, that cannot be determined beforehand because all non-explicit rules collide with each other (just create the union of both froms). To maybe avoid some non-explicit rule collisions, you could add another layer of priority namely minimizing the To size (going for the greatest reduction if there are multiple matching non-explicit rules of the same size). But first I'd do it without that and then add it if necessary (don't forget to comment).